refactor: move dead full-transcript engine projector to test-only
Engine.ChatSnapshot/scanPersistedTranscript were dead in production (guard-blocked)
but compiled a full events.jsonl WalkEvents into the binary; TranscriptPageSnapshot
had no callers at all. Delete TranscriptPageSnapshot, relocate ChatSnapshot +
scanPersistedTranscript into a runtime _test.go (in-package test helper), and
migrate the three cross-package test callers to the bounded RecentTailTranscriptWindow.
